% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMDM_expression.R
\docType{data}
\name{BMDM_expression}
\alias{BMDM_expression}
\title{Gene expression of BMDM cells}
\format{
A data frame with 34400 rows and 12 variables
\describe{
  \item{Count_0h_A}{control condition, rep A}
  \item{Count_0h_B}{control condition, rep B}
  \item{Count_0hLPS_A}{LPS condition, rep A}
  \item{Count_0hLPS_B}{LPS condition, rep B}
  \item{Count_9h_A}{pH condition, rep A}
  \item{Count_9h_B}{pH condition, rep B}
  \item{Count_5hLPS_A}{pH & LPS condition, rep A}
  \item{Count_5hLPS_B}{pH & LPS condition, rep B}
  \item{q_Ph0hLPS}{q stat}
  \item{q_Ph5hLPS}{q stat}
  \item{q_LPS5hvs0h}{q stat}
  \item{q_Ph9h}{q stat}
}
}
\source{
in house data set
}
\usage{
data(BMDM_expression)
}
\description{
A dataset containing gene expression count data from Bone Marrow Derived Macrophages
after being subjected to two signals, pH-6.4 and LPS. Contains two replicates, A and B
}
\keyword{datasets}
