% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ratios.R
\name{compute_ratios}
\alias{compute_ratios}
\title{Compute log2 ratios of 2 signals}
\usage{
compute_ratios(data, pseudo = 1, lowess_norm = 1)
}
\arguments{
\item{data}{Data should be formatted with every row being a gene and every column being an experimental condition}

\item{pseudo}{Add a pseudo count in order to deal with zero counts. Default is 1}

\item{lowess_norm}{Uses lowess normalization in order to improve data quality. It is performed by default}
}
\description{
The function will ask you to enter column numbers per condition of the experiment
If you have multiple replicates, enter the column numbers with a comma inbetween.
For example: control condition is in columns 1 and 2. Enter: 1,2
The output of the function is a dataframe with LogRatios to be used for the fitting of the model
}
\details{
`compute_ratios` returns the log2 ratios with option of lowess normalization
}
