% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_standardization.R
\name{geomean}
\alias{geomean}
\title{Geometric mean of a vector}
\usage{
geomean(x)
}
\arguments{
\item{x}{should be a numeric vector}
}
\description{
`geomean` returns the geometric mean of a vector x
}
