% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{ratios.fit}
\alias{ratios.fit}
\title{Fits ratios to a linear model}
\usage{
ratios.fit(ratios, CompThreshold = 1.5, n_rep = 2)
}
\arguments{
\item{ratios}{input ratios data}

\item{CompThreshold}{ratio threshold for significance}

\item{n_rep}{number of replicates in data}
}
\description{
`ratios.fit` ratios are fitted to a linear model
}
