% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ratios.R
\name{RNAseqLowess}
\alias{RNAseqLowess}
\title{LOWESS normalization of LogRatios}
\usage{
RNAseqLowess(LogIntensity, LogRatios)
}
\arguments{
\item{LogIntensity}{Uses the LogIntensity computed by the compute_ratios function}

\item{LogRatios}{Uses the LogRatios computed by the compute_ratios function}
}
\description{
The function uses the loess function from R with the following parameters. span=0.05, degree=1, family=gaussian and iter=1
}
\details{
`RNAseqLowess` returns the lowess normalized logratios
}
