% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_standardization.R
\name{MedianNorm}
\alias{MedianNorm}
\title{Median Normalization of the data}
\usage{
MedianNorm(data, countthres = 10, pseudo = 1)
}
\arguments{
\item{data}{Use a count matrix K_ij, with one row for each gene i and one column for each sample j. Matrix entries indicate the number of sequencing reads mapped to a gene in a sample.}

\item{countthres}{Count threshold for the count data. Default is 10}

\item{pseudo}{Pseudo count added to deal with zero count problem. Default is 1}
}
\description{
`MedianNorm` normalizes data with the median normalization method
}
